feat: Removes WebBase (reusing ArcBase for both WebBase and MapBase)
