+++ src/ts/Test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import '../css/index.scss';
-import { Map, Scene } from './components/ArcComposites';
+import { Map, Scene } from './components/MapComposites';
 import { WebMap, WebScene } from './components/WebComposites';
 import { BasemapToggle } from './components/widgets/WidgetComposites';
 import { ElevationLayer } from './components/layers/LayerComposites';
@@ -12,24 +12,31 @@ export default class TestComponent extends React.Component<null, ComponentState>
 
     public render() {
         return (
-            <Scene
-                className="half-map"
-                mapProperties={{ basemap: 'topo', ground: 'world-elevation' }}
-                viewProperties={{
-                    camera: {
-                        heading: 300,
-                        position: [-121.83, 48.279, 1346],
-                        tilt: 60
-                    }
-                }}
-            >
-                <ElevationLayer
-                    onLayerviewCreate={() => console.log('Created the elevation layer.')}
-                    layerProperties={{
-                        url: 'https://sampleserver6.arcgisonline.com/arcgis/rest/services/OsoLandslide/OsoLandslide_After_3DTerrain/ImageServer'
+            <div>
+                <Scene
+                    className="half-map"
+                    mapProperties={{ basemap: 'topo', ground: 'world-elevation' }}
+                    viewProperties={{
+                        camera: {
+                            heading: 300,
+                            position: [-121.83, 48.279, 1346],
+                            tilt: 60
+                        }
                     }}
+                >
+                    <ElevationLayer
+                        onLayerviewCreate={() => console.log('Created the elevation layer.')}
+                        layerProperties={{
+                            url: 'https://sampleserver6.arcgisonline.com/arcgis/rest/services/OsoLandslide/OsoLandslide_After_3DTerrain/ImageServer'
+                        }}
+                    />
+                </Scene>
+                <WebMap
+                    id="6627e1dd5f594160ac60f9dfc411673f"
+                    className="half-map"
+                    onLoad={(map, view) => { console.log(map, view); }}
                 />
-            </Scene>
+            </div>
         );
     }
 }
