+++ src/ts/components/WebComposites.tsx
@@ -1,17 +1,20 @@
 import * as React from 'react';
-import { WebView, BaseProps } from './WebBase';
+import { BaseProps } from './ArcBase';
+import { WebBase } from './ArcComposites';
 
 interface MapProps extends BaseProps {
+    id: string;
     viewProperties?: __esri.WebMapProperties;
 }
 
 interface SceneProps extends BaseProps {
+    id: string;
     viewProperties?: __esri.WebSceneProperties;
 }
 
 export const WebMap = (props: MapProps) => (
-    <WebView
-        scriptUri={['esri/WebMap', 'esri/views/MapView']}
+    <WebBase
+        scriptUri={['esri/WebMap', 'esri/views/MapView', 'dojo/promise/all']}
         {...props}
         viewProperties = {{...props.viewProperties}}
         mapProperties = {{...props.mapProperties}}
@@ -19,8 +22,8 @@ export const WebMap = (props: MapProps) => (
 );
 
 export const WebScene = (props: SceneProps) => (
-    <WebView
-        scriptUri={['esri/WebScene', 'esri/views/SceneView']}
+    <WebBase
+        scriptUri={['esri/WebScene', 'esri/views/SceneView', 'dojo/promise/all']}
         {...props}
         viewProperties = {{...props.viewProperties}}
         mapProperties = {{...props.mapProperties}}
